Inline Case files for LoopWithTooManyJumpStatementsSpec
